+++ ErsatzTV.FFmpeg/Format/AudioLayout.cs
@@ -1,0 +1,6 @@
+namespace ErsatzTV.FFmpeg.Format;
+
+public static class AudioLayout
+{
+    public const string Stereo = "stereo";
+}